Footer and State fixes
The Markdown object was being reused in an attempt to be efficient. Unfortunately, the footnotes modules alters the state of the Markdown object and thus footers were being put on pages that didn't have any.

Also, the css for the footer has been modified to make it 'sticky'. It should now be at the bottom of the window at a minimum.